Restructure the project around a bit.
